+++ src/test/java/racingGame/RandomScoreGeneratorTest.java
@@ -5,17 +5,17 @@ import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Random;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
-class ScoreGeneratorTest {
+class RandomScoreGeneratorTest {
 
     @Test
     @DisplayName("동일 값 출력 확인")
     void equalityTest() {
-        ScoreGenerator scoreGenerator = new ScoreGenerator(0);
-        ScoreGenerator scoreGenerator1 = new ScoreGenerator(0);
+        RandomScoreGenerator randomScoreGenerator = new RandomScoreGenerator(0);
+        RandomScoreGenerator randomScoreGenerator1 = new RandomScoreGenerator(0);
 
         int numTry = 10;
 
@@ -23,8 +23,8 @@ class ScoreGeneratorTest {
         List<Integer> result1 = new ArrayList<>();
 
         for (int i = 0; i < numTry; i++) {
-            result.add(scoreGenerator.generateScore());
-            result1.add(scoreGenerator1.generateScore());
+            result.add(randomScoreGenerator.generateScore());
+            result1.add(randomScoreGenerator1.generateScore());
         }
 
         assertThat(result).isEqualTo(result1);
@@ -35,8 +35,8 @@ class ScoreGeneratorTest {
     @DisplayName("매번 다른 값 출력하는 지 테스트")
     void diverseOutput() {
 
-        ScoreGenerator scoreGenerator = new ScoreGenerator();
-        ScoreGenerator scoreGenerator1 = new ScoreGenerator();
+        RandomScoreGenerator randomScoreGenerator = new RandomScoreGenerator();
+        RandomScoreGenerator randomScoreGenerator1 = new RandomScoreGenerator();
 
         int numTry = 100;
 
@@ -44,8 +44,8 @@ class ScoreGeneratorTest {
         List<Integer> result1 = new ArrayList<>();
 
         for (int i = 0; i < numTry; i++) {
-            result.add(scoreGenerator.generateScore());
-            result1.add(scoreGenerator1.generateScore());
+            result.add(randomScoreGenerator.generateScore());
+            result1.add(randomScoreGenerator1.generateScore());
         }
 
         assertThat(result).isNotEqualTo(result1);
